refactor: runner timeout code
