+++ pkg/runner/step.go
@@ -10,6 +10,7 @@ import (
 	"github.com/kyverno/chainsaw/pkg/resource"
 	"github.com/kyverno/chainsaw/pkg/runner/logging"
 	"github.com/kyverno/chainsaw/pkg/runner/operations"
+	"github.com/kyverno/chainsaw/pkg/runner/timeout"
 	"github.com/kyverno/kyverno/ext/output/color"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	ctrlclient "sigs.k8s.io/controller-runtime/pkg/client"
@@ -65,7 +66,7 @@ func executeStep(t *testing.T, logger logging.Logger, ctx Context, basePath stri
 			if err := ctx.namespacer.Apply(&resource); err != nil {
 				fail(t, operation.ContinueOnError)
 			}
-			operationCtx, cancel := timeoutCtx(DefaultDeleteTimeout, config.Timeouts.Delete, test.Timeouts.Delete, step.Spec.Timeouts.Delete, operation.Timeout)
+			operationCtx, cancel := timeout.Context(timeout.DefaultDeleteTimeout, config.Timeouts.Delete, test.Timeouts.Delete, step.Spec.Timeouts.Delete, operation.Timeout)
 			defer cancel()
 			if err := operations.Delete(operationCtx, logger, &resource, c); err != nil {
 				fail(t, operation.ContinueOnError)
@@ -76,7 +77,7 @@ func executeStep(t *testing.T, logger logging.Logger, ctx Context, basePath stri
 	if !skipDelete(config.SkipDelete, test.SkipDelete, step.Spec.SkipDelete) {
 		cleanup = func(obj ctrlclient.Object, c client.Client) {
 			t.Cleanup(func() {
-				cleanupCtx, cancel := timeoutCtx(DefaultCleanupTimeout, config.Timeouts.Cleanup, test.Timeouts.Cleanup, step.Spec.Timeouts.Cleanup, nil)
+				cleanupCtx, cancel := timeout.Context(timeout.DefaultCleanupTimeout, config.Timeouts.Cleanup, test.Timeouts.Cleanup, step.Spec.Timeouts.Cleanup, nil)
 				defer cancel()
 				if err := operations.Delete(cleanupCtx, logger, obj, c); err != nil {
 					t.Fail()
@@ -86,7 +87,7 @@ func executeStep(t *testing.T, logger logging.Logger, ctx Context, basePath stri
 	}
 	for _, operation := range step.Spec.Exec {
 		func() {
-			operationCtx, cancel := timeoutCtx(DefaultExecTimeout, config.Timeouts.Exec, test.Timeouts.Exec, step.Spec.Timeouts.Exec, operation.Timeout)
+			operationCtx, cancel := timeout.Context(timeout.DefaultExecTimeout, config.Timeouts.Exec, test.Timeouts.Exec, step.Spec.Timeouts.Exec, operation.Timeout)
 			defer cancel()
 			if err := operations.Exec(operationCtx, logger, operation.Exec, !operation.SkipLogOutput, ctx.namespacer.GetNamespace()); err != nil {
 				fail(t, operation.ContinueOnError)
@@ -107,7 +108,7 @@ func executeStep(t *testing.T, logger logging.Logger, ctx Context, basePath stri
 					logger.Log("LOAD  ", color.BoldRed, err)
 					fail(t, operation.ContinueOnError)
 				}
-				operationCtx, cancel := timeoutCtx(DefaultApplyTimeout, config.Timeouts.Apply, test.Timeouts.Apply, step.Spec.Timeouts.Apply, operation.Timeout)
+				operationCtx, cancel := timeout.Context(timeout.DefaultApplyTimeout, config.Timeouts.Apply, test.Timeouts.Apply, step.Spec.Timeouts.Apply, operation.Timeout)
 				defer cancel()
 				if err := operations.Apply(operationCtx, logger, resource, c, shouldFail, cleanup); err != nil {
 					fail(t, operation.ContinueOnError)
@@ -128,7 +129,7 @@ func executeStep(t *testing.T, logger logging.Logger, ctx Context, basePath stri
 					logger.Log("LOAD  ", color.BoldRed, err)
 					fail(t, operation.ContinueOnError)
 				}
-				operationCtx, cancel := timeoutCtx(DefaultAssertTimeout, config.Timeouts.Assert, test.Timeouts.Assert, step.Spec.Timeouts.Assert, operation.Timeout)
+				operationCtx, cancel := timeout.Context(timeout.DefaultAssertTimeout, config.Timeouts.Assert, test.Timeouts.Assert, step.Spec.Timeouts.Assert, operation.Timeout)
 				defer cancel()
 				if err := operations.Assert(operationCtx, logger, resources[i], c); err != nil {
 					fail(t, operation.ContinueOnError)
@@ -149,7 +150,7 @@ func executeStep(t *testing.T, logger logging.Logger, ctx Context, basePath stri
 					logger.Log("LOAD  ", color.BoldRed, err)
 					fail(t, operation.ContinueOnError)
 				}
-				operationCtx, cancel := timeoutCtx(DefaultErrorTimeout, config.Timeouts.Error, test.Timeouts.Error, step.Spec.Timeouts.Error, operation.Timeout)
+				operationCtx, cancel := timeout.Context(timeout.DefaultErrorTimeout, config.Timeouts.Error, test.Timeouts.Error, step.Spec.Timeouts.Error, operation.Timeout)
 				defer cancel()
 				if err := operations.Error(operationCtx, logger, resources[i], c); err != nil {
 					fail(t, operation.ContinueOnError)
